Backend/requirements.txt
fastapi
uvicorn
pdfplumber
camelot-py[cv]
python-docx
pytesseract
layoutparser
pandas
streamlit
